disallow core::time::Duration (not only) for sleep API (#5802)
* disallow core::time::Duration (not only) for sleep API

* fmt run

+++ esp-hal/src/rtc_cntl/sleep/esp32c3.rs
@@ -97,7 +97,7 @@ impl WakeSource for TimerWakeupSource {
     ) {
         triggers.set_timer(true);
         // TODO: maybe add check to prevent overflow?
-        let ticks = crate::clock::us_to_rtc_ticks(self.duration.as_micros() as u64);
+        let ticks = crate::clock::us_to_rtc_ticks(self.duration.as_micros());
         // "alarm" time in slow rtc ticks
         let now = rtc.time_since_boot_raw();
         let time_in_ticks = now + ticks;
